alpha/submitted-dl-support.diff
older versions of glibc would build dl-sysdep as shared-only and dl-support as
static-only.  alpha hooks in a cache variable via dl-auxv.h.  newer versions of
glibc build dl-sysdep as both shared and static which means we now have symbol
duplication for static builds with dl-sysdep and dl-support.  since dl-sysdep
is both shared/static, there is no point in hooking dl-support anymore, so we
can punt it.